[NC-129] add support Token-based authentication
Problem: the HasClient instance for Macaroon conflicts with the one
in Servant.Auth.Client.
The instance is implicitly exported by other modules (most notably:
Lightning.Node.Api) and as a result it's not possible to make a
servant client for a service that uses this library and has a
different authentication method (e.g. JWT).

Solution: add an implementation of 'HasClient' that supports both
Token-based and Macaroon-based authentication.
Additionally exports copies of Token and Bearer in order to cover
the functionality of Servant.Auth.Client